+++ NEWS.md
@@ -26,5 +26,7 @@ Moved arrow labels toward the end of the arrows.
 flowSet arguments are non-null, distances are calculated for all sample pairs, 
 where the first element comes from `fs`, 
 and the second element comes from `fs2`.
-- renamed `ggplotSamplesMDS` into `ggplotSampleMDS` and 
-`ggplotSamplesMDSShepard` into `ggplotSampleMDSShepard`
+- renamed `ggplotSamplesMDS` into `ggplotSampleMDS`
+- renamed `ggplotSamplesMDSShepard` into `ggplotSampleMDSShepard`
+- renamed `getChannelsSummaryStat` into `getChannelSummaryStats`
+- new function `ggplotSampleMDSWrapBiplots()`